Restaurant/requirements.txt
asgiref==3.8.1
Django==5.0.9
django-cors-headers==4.4.0
djangorestframework==3.15.2
gunicorn==23.0.0
mssql-django==1.5
packaging==24.1
pyodbc==5.1.0
pytz==2024.2
sqlparse==0.5.1
tzdata==2024.2
